+++ checker/tests/custom-alias/NonNull.java
@@ -1,0 +1,7 @@
+package custom.alias;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE_USE)
+public @interface NonNull {}