+++ checker/tests/custom-alias/Pure.java
@@ -1,0 +1,3 @@
+package custom.alias;
+
+public @interface Pure {}