+++ framework/src/main/java/org/checkerframework/framework/source/SourceChecker.java
@@ -254,6 +254,18 @@ import io.github.classgraph.ClassGraph;
     "ignoreInvalidAnnotationLocations",
 
     ///
+    /// Compatibility options
+    ///
+
+    // Additional type and declaration annotation aliases
+    // -AaliasedTypeAnnos={aliases} or -AaliasedDeclAnnos={aliases}
+    // where `aliases` is in the format
+    // `FQN.canonical.Qualifier1:FQN.alias1.Qual1,FQN.alias2.Qual1;FQN.canonical.Qualifier2:FQN.alias1.Qual2`
+    // org.checkerframework.framework.type.AnnotatedTypeFactory
+    "aliasedTypeAnnos",
+    "aliasedDeclAnnos",
+
+    ///
     /// Partially-annotated libraries
     ///